requirements.txt
fastapi
uvicorn
python-telegram-bot==20.7
psycopg2-binary
requests
openai
google-api-python-client
sqlalchemy
reportlab